Refactors code  - Adds unit testing  - Converts model to kt
