+++ data/src/main/java/com/globant/equattrocchio/data/ImageRepository.kt
@@ -1,13 +1,13 @@
 package com.globant.equattrocchio.data
 
 import android.util.Log
-import com.globant.equattrocchio.data.service.api.ImagesApi
+import com.globant.equattrocchio.data.service.api.ImagesApiImpl
 import com.globant.equattrocchio.domain.models.ResultDomainInput
 import com.globant.equattrocchio.domain.service.ImagesServices
 import io.reactivex.Observable
 import javax.inject.Inject
 
-class ImagesRepository @Inject constructor(private val mapper: ImageMapper, private val imagesApi: ImagesApi) : ImagesServices {
+class ImagesRepository @Inject constructor(private val mapper: ImageMapper, private val imagesApi: ImagesApiImpl) : ImagesServices {
 
 
     override fun getLatestImagesRx(): Observable<ResultDomainInput> {
