--- data/src/main/java/com/globant/equattrocchio/data/response/Image.java
@@ -1,52 +1,0 @@
-package com.globant.equattrocchio.data.response;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Image {
-    @SerializedName("id")
-    @Expose
-    private Integer id;
-    @SerializedName("url")
-    @Expose
-    private String url;
-    @SerializedName("large_url")
-    @Expose
-    private String largeUrl;
-    @SerializedName("source_id")
-    @Expose
-    private Object sourceId;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public void setUrl(String url) {
-        this.url = url;
-    }
-
-    public String getLargeUrl() {
-        return largeUrl;
-    }
-
-    public void setLargeUrl(String largeUrl) {
-        this.largeUrl = largeUrl;
-    }
-
-    public Object getSourceId() {
-        return sourceId;
-    }
-
-    public void setSourceId(Object sourceId) {
-        this.sourceId = sourceId;
-    }
-
-}
