+++ data/src/main/java/com/globant/equattrocchio/data/response/ImageCache.kt
@@ -1,0 +1,3 @@
+package com.globant.equattrocchio.data.response
+
+data class Image(var id: Int = 0, var url: String = "", var largeUrl: String = "", var sourceId: Any = "")
