+++ data/src/main/java/com/globant/equattrocchio/data/response/ResultDataInput.kt
@@ -1,0 +1,3 @@
+package com.globant.equattrocchio.data.response
+
+data class ResultInput(var images: List<Image> = emptyList())
