--- data/src/main/java/com/globant/equattrocchio/data/response/ResultInput.java
@@ -1,21 +1,0 @@
-package com.globant.equattrocchio.data.response;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-import java.util.List;
-
-public class ResultInput {
-    @SerializedName("images")
-    @Expose
-    private List<Image> mImages = null;
-
-    public List<Image> getImages() {
-        return mImages;
-    }
-
-    public void setImages(List<Image> images) {
-        this.mImages = images;
-    }
-
-}
