--- data/src/main/java/com/globant/equattrocchio/data/service/api/IImagesApi.kt
@@ -1,5 +1,0 @@
-package com.globant.equattrocchio.data.service.api
-
-interface IImagesApi {
-    fun getInstance(): SplashbaseApi
-}
