--- domain/src/main/java/com/globant/equattrocchio/domain/UseCase.java
@@ -1,48 +1,0 @@
-package com.globant.equattrocchio.domain;
-
-import com.globant.equattrocchio.domain.models.ResultDomainInput;
-
-import io.reactivex.Observable;
-import io.reactivex.disposables.CompositeDisposable;
-import io.reactivex.disposables.Disposable;
-import io.reactivex.observers.DisposableObserver;
-import io.reactivex.schedulers.Schedulers;
-
-public abstract class UseCase<T, Params> {
-
-    private final CompositeDisposable disposables;
-
-    UseCase() {
-        this.disposables = new CompositeDisposable();
-    }
-
-    abstract Observable<ResultDomainInput> buildUseCaseRx();
-
-    public void execute(DisposableObserver<T> observer, Params params) {
-        Preconditions.checkNotNull(observer);
-        addDisposable(observer);
-    }
-
-    public Observable<ResultDomainInput> executeRx() {
-        return this.buildUseCaseRx()
-                .subscribeOn(Schedulers.io());
-    }
-
-    /**
-     * Dispose from current {@link CompositeDisposable}.
-     */
-    public void dispose() {
-        if (!disposables.isDisposed()) {
-            disposables.dispose();
-        }
-    }
-
-    /**
-     * Dispose from current {@link CompositeDisposable}.
-     */
-    private void addDisposable(Disposable disposable) {
-        Preconditions.checkNotNull(disposable);
-        Preconditions.checkNotNull(disposables);
-        disposables.add(disposable);
-    }
-}
