+++ domain/src/main/java/com/globant/equattrocchio/domain/models/Image.kt
@@ -1,0 +1,4 @@
+package com.globant.equattrocchio.domain.models
+
+class Image {
+}
