+++ domain/src/main/java/com/globant/equattrocchio/domain/models/ResultDomainInput.kt
@@ -1,0 +1,4 @@
+package com.globant.equattrocchio.domain.models
+
+class ResultDomainInputs {
+}
