--- presentation/src/main/java/com/globant/equattrocchio/cleanarchitecture/Info.kt
@@ -1,9 +1,0 @@
-package com.globant.equattrocchio.cleanarchitecture
-
-import javax.inject.Inject
-import javax.inject.Singleton
-
-@Singleton
-class Info @Inject constructor() {
-    var text = "Hello"
-}
