+++ presentation/src/main/java/com/globant/equattrocchio/cleanarchitecture/di/module/SubjectsModule.kt
@@ -1,0 +1,4 @@
+package com.globant.equattrocchio.cleanarchitecture.di.module
+
+class SubjectsModule {
+}
