+++ presentation/src/main/java/com/globant/equattrocchio/cleanarchitecture/utils/Constants.kt
@@ -1,0 +1,4 @@
+package com.globant.equattrocchio.cleanarchitecture.utils
+
+object Constants {
+}
